Cell highlighting feature (#25)
* Adds cell highlighting feature

Cells can now be highlighted bright red. Highlighting mode is toggled by pressing H. When highlighting mode is activated, the currently selected cell(s) are highlighted. If a single cell is highlighted, the currently highlighted cell can be selected with the + and - buttons, which cycles which cell is currently being highlighted. The highlighted cell(s) along with whether highlighting is enabled is displayed in text labels in the top right of the screen. Up to two cells can be selected at a time. Also, clicking an empty region of the frame now has the equivalent effect as pressing escape and deselects any selected cells.

* Fixes highlighting bug

In previous commit, it was possible to highlight a cell ID that did not actually exist in the file. This problem was caused by the num_tracks variable was incorrectly incremented upon the file being opened. This commit also fixes a glitch with the contrast adjustment feature, which also relied on the num_tracks variable, which previously permitted the contrast to be adjusted too far which would cause all the cells to be rendered black.

* minor edits for clarity

-comments mark the code sections that change selection behavior
(ie, when background is clicked, deselect cells)
-no need to deselect when self.mode.kind is None and background is clicked