Made rendering speed-up changes, added training loss plot, fixed hover edge bug, made plotly adjustment, separated training and prediction logic
Relevant canvases now use a smaller cell matrix and use the actual list of cells on a given frame. A plot that tracks training loss is now created upon training, and training does not auto-predict (instead, a predict button is used to perform predictions on a given frame). Additional UI changes such as the hover tooltip persisting on edges and plotly settings were made.